Cleaned up code and switched over to HTTPClient
Big cleanup, only relying on HTTPClient for transmitting data now. This way we don't need to
worry about parsing headers and whether it is an HTTP or HTTPS connection. The code will
figure out if SSL is needed and will load the required root_ca.pem from the SPIFFS filesystem.